Move hook to user lib
